Add floating point support

+++ benchmarks/dynamatic/gemm/gemm.h
@@ -1,5 +1,5 @@
-typedef float in_float_t;
-typedef float out_float_t;
-typedef float inout_float_t;
+typedef int in_float_t;
+typedef int out_float_t;
+typedef int inout_float_t;
 
 int gemm(in_float_t alpha, in_float_t beta, in_float_t A[30][30], in_float_t B[30][30], inout_float_t C[30][30] );
