+++ benchmarks/dynamatic/img_avg/img_avg.h
@@ -1,6 +1,6 @@
-typedef float in_float_t;
-typedef float out_float_t;
-typedef float inout_float_t;
+typedef int in_float_t;
+typedef int out_float_t;
+typedef int inout_float_t;
 
 typedef int in_int_t;
 typedef int inout_int_t;
